# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AI Alpha API",
    "description": "API documentation for AI Alpha",
    "version": "2.1"
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/chart/ohlc": {
      "get": {
        "tags": [
          "Chart"
        ],
        "summary": "Retrieve ohlc (open, high, low, close) chart data for a specified cryptocurrency.",
        "description": "Retrieve ohlc (open, high, low, close) chart data for a specified cryptocurrency.",
        "parameters": [
          {
            "name": "gecko_id",
            "in": "query",
            "description": "The CoinGecko ID of the cryptocurrency",
            "required": true,
            "type": "string",
            "schema": {}
          },
          {
            "name": "vs_currency",
            "in": "query",
            "description": "The target currency for market data",
            "required": false,
            "type": "string",
            "schema": {}
          },
          {
            "name": "interval",
            "in": "query",
            "description": "The time interval for OHLC data",
            "required": false,
            "type": "string",
            "schema": {}
          },
          {
            "name": "precision",
            "in": "query",
            "description": "The number of decimal places for price data",
            "required": false,
            "type": "integer",
            "schema": {}
          },
          {
            "name": "symbol",
            "in": "query",
            "description": "The trading symbol of the cryptocurrency",
            "required": true,
            "type": "string",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved OHLC data",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "minItems": 5,
                    "maxItems": 5
                  },
                  "description": "Array of OHLC data points [timestamp, open, high, low, close]"
                },
                "error": {
                  "type": "string",
                  "nullable": true
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "null"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "null"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}